course chapters initial
